+++ src/taskgraph/parameters.py
@@ -32,6 +32,7 @@ base_schema = Schema(
     {
         Required("base_repository"): str,
         Required("base_ref"): str,
+        Required("base_rev"): str,
         Required("build_date"): int,
         Required("build_number"): int,
         Required("do_not_optimize"): [str],
@@ -84,9 +85,11 @@ def _get_defaults(repo_root=None):
         repo_url = ""
         project = ""
 
+    default_base_ref = repo.default_branch
     return {
         "base_repository": repo_url,
-        "base_ref": repo.default_branch,
+        "base_ref": default_base_ref,
+        "base_rev": repo.find_latest_common_revision(default_base_ref, repo.head_rev),
         "build_date": int(time.time()),
         "build_number": 1,
         "do_not_optimize": [],